Added: task (7. Average score) for 0x00 MySQL Advanced

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -2,7 +2,7 @@
 DELIMITER //
 CREATE PROCEDURE ComputeAverageScoreForUser(IN user_id INT)
 BEGIN
-  DECLARE average INT;
+  DECLARE average FLOAT;
   SELECT AVG(score) INTO average
     FROM corrections
     WHERE corrections.user_id = user_id
